added nice modular code :)
