--- extra_types.py
@@ -1,2 +1,0 @@
-from typing import List, Tuple
-type PSRSequence = List[Tuple[List[int], List[int], List[int]]]
