--- logs/training_20240923_211530.log
@@ -1,4 +1,0 @@
-2024-09-23 21:15:30,938 - INFO - Starting main execution
-2024-09-23 21:15:30,939 - INFO - Loading data from 23-24-problem_logs.csv
-2024-09-23 21:15:38,954 - INFO - Data loaded. Shape: (12075734, 9)
-2024-09-23 21:15:38,954 - INFO - Processing data
