--- logs/training_20240923_222407.log
@@ -1,15 +1,0 @@
-2024-09-23 22:24:07,025 - INFO - Starting main execution
-2024-09-23 22:24:07,025 - INFO - Loading data from 23-24-problem_logs.csv
-2024-09-23 22:24:14,608 - INFO - Data loaded. Shape: (12075734, 9)
-2024-09-23 22:24:14,608 - INFO - Processing data
-2024-09-23 22:25:58,520 - INFO - Data processing completed. Time taken: 111.49 seconds
-2024-09-23 22:25:58,520 - INFO - Creating skill and problem maps
-2024-09-23 22:25:58,638 - INFO - Maps created. Unique skills: 525, Unique problems: 90065
-2024-09-23 22:25:58,638 - INFO - Time taken: 0.12 seconds
-2024-09-23 22:25:58,640 - INFO - Creating data generators
-2024-09-23 22:25:58,640 - INFO - Creating model
-2024-09-23 22:25:58,690 - INFO - Starting model training
-2024-09-23 22:25:58,706 - INFO - Starting model training
-2024-09-23 22:25:58,706 - INFO - Model configuration: batch_size=32, accumulation_steps=4
-2024-09-23 22:25:58,709 - INFO - Optimizer: Adam, Learning rate: 0.0010000000474974513
-2024-09-23 22:25:58,711 - INFO - Starting epoch 1/10
