--- logs/training_20240923_224309.log
@@ -1,2 +1,0 @@
-2024-09-23 22:43:09,916 - INFO - Starting main execution
-2024-09-23 22:43:09,916 - INFO - Loading data from 23-24-problem_logs.csv
